Handle test resources service lifecycle
This commit reworks how the test resources service is started
and shutdown, to mimic what is done in the Gradle plugin:

- if `mn:start-testresources-service` is explicitly called, then
the service will be started in the background: it's then the
responsibility of the user to stop it using `mn:stop-testresources-service`
- otherwise, the test resources service is started before tests
are executed, and shutdown after
- in addition, the `mn:run` task will also start/shutdown the service
accordingly

Note that in order to be able to shutdown the test resources service,
we need to figure out how to contact it, and this is done by reading
the server settings properties file. However, if we kept that file
in the regular `target` build directory, then any call to `mvn clean`
would delete that file, which would have a couple of side effects:

- the background test resources service would never be shutdown
- a new test resources service would be started each time `clean`
is called

To workaround this, the settings file is written _outside_ the build
directory, in `.micronaut/test-resources` (in the project folder).