Replace openjdk with eclipse-temurin (#647)
* Replace openjdk with eclipse-temurin

The `openjdk` project has been deprecated and only publishes early access
builds[1]. As a result the `openjdk:17-alpine` on Dockerhub has not been
updated in the last year and a half.

Replacing the image with `eclipse-temurin` should resolve these issues. The
choice for `eclipse-temurin` is somewhat arbitrary. But as an opensource
project its good enough.

 1. https://hub.docker.com/_/openjdk

Fixes: #628

# Conflicts:
#	micronaut-maven-jib-integration/src/main/java/io/micronaut/maven/jib/JibMicronautExtension.java

* Replace openjdk with eclipse-temurin

The `openjdk` project has been deprecated and only publishes early access
builds[1]. As a result the `openjdk:17-alpine` on Dockerhub has not been
updated in the last year and a half.

Replacing the image with `eclipse-temurin` should resolve these issues. The
choice for `eclipse-temurin` is somewhat arbitrary. But as an opensource
project its good enough.

 1. https://hub.docker.com/_/openjdk

Fixes: #628

---------

Co-authored-by: M.P. Korstanje <rien.korstanje@gmail.com>