Initial implementation of OpenAPI generation (#747)
* Initial implementation of OpenAPI generation

This commit is the equivalent of https://github.com/micronaut-projects/micronaut-gradle-plugin/pull/736.
It provides a couple of mojos which are responsible for generating either a client or a server
from an OpenAPI definition file.

This implementation will generate the code in the `target/generated-sources` directory.
For now, generation is limited to API and model classes (no test sources generation, which would
most likely live under Launch if ever added).

* Upgrade to Micronaut OpenAPI Generator 5.0.0-M6

* Add documentation for the OpenAPI integration

* Fix spacing [skip ci]

---------

Co-authored-by: Álvaro Sánchez-Mariscal <alvaro.sanchez-mariscal@oracle.com>