Add a generic mojo for OpenAPI generation
This upgrades the Maven plugin to the latest Micronaut OpenAPI milestone.
It adds a new generic mojo which can be used for cases where the user needs
to override the default Micronaut generators. This works by adding a new
configuration property which provides the name of the generator class, and
adding the generator dependency as a dependency of the Micronaut Maven plugin
(as demonstrated in the integration test).

In order to be able to configure the generator, we create the configuration
builder, then conventionally map properties from the plugin configuration
to builder methods. For example, for the following configuration:

```xml
    <configuration>
        <properties>
            <micronaut.openapi.generator.properties.greeting>Hello, world!</micronaut.openapi.generator.properties.greeting>
            <micronaut.openapi.generator.properties.joy>true</micronaut.openapi.generator.properties.joy>
            <micronaut.openapi.generator.properties.magicNumber>123</micronaut.openapi.generator.properties.magicNumber>
        </properties>
    </configuration>
```

We automatically invoke the `withGreeting`, `withJoy` and `withMagicNumber` methods.
It's worth noting that currently we only support 3 types: String, boolean and int.
Should more types be supported, we'd have to extend the dynamic invoker.