Fix CRaC JDK download in docker container
Azul added a JRE to their CRaC java deliverables.

The query wasn't specific enough to say we want a JDK, so we end up pulling the JRE if it comes back first in the response.

This change adds java_package_type=jdk to the request URL so we just get the JDK.