Use an Arm compatible base image for docker-native on arm (#847)
* Use an Arm compatible base image for docker-native on arm

For both Arm and X86 we are using an alpine base image.

There is no Arm variant of Alpine, so this builds an ARM native binary, then puts it in an x86 base image.

This change switches to use cgr.dev/chainguard/wolfi-base:latest when the architecture is Arm.

This image was chosen due to it's size... There may be a better default choice out there...

* Remove logging which was causing a test to fail

* Apply suggestions from code review

Co-authored-by: Álvaro Sánchez-Mariscal <alvaro.sanchez-mariscal@oracle.com>

* Test ARM Dockerfile on CI (which is x86)

---------

Co-authored-by: Álvaro Sánchez-Mariscal <alvaro.sanchez-mariscal@oracle.com>