Fix test resources service not being shutdown
This commit fixes the fact that the test resources service could be
kept alive even when invoking `mn:stop-testresources-service`. This
was caused by the presence of a stale keepalive file. The scenario
to trigger this was:

1. execute `mvn mn:start-testresources-service`
2. execute `mvn run`
3. when the Micronaut server is running, interrupt with CTRL+C
4. execute `mvn mn:stop-testresources-service`

In that case the server was kept alive because the keepalive
file was not deleted.