Fix test resources service keepalive (#910)
This commit fixes the logic for shutting down the test resources service.
Before, the "keepAlive" flag was exposed to the user. If it was set to
true, then invoking `stop-testresources-service` would always keep the
service alive. Now, the flag is purely internal and used to keep track
of whether the service was _already_ alive before starting the build
or not.

The keepalive file is now kept under the temporary directory: this
prevents the file from being accidentally deleted by another process,
or even conflicts with different namespaces for shared services.

If the service was alive before running the build, it will be kept
alive unless `stop-testresources-service` is called. If it wasn't
alive, then it will be shutdown at the end of the build.

The logic is now similar to what the Gradle plugin is doing.

In addition, this commit adds support for the `debug-server` property
which allows running the test resources service in debug mode.

The fix will only be complete once https://github.com/micronaut-projects/micronaut-test-resources/pull/464
is merged.