Fix race conditions in watch mode
This commit fixes the watch mode for several issues, which are best
highlighted by a long application startup:

- there can now only be a single compilation request at a time. Previously,
if compilation took more time than the time it takes to get another watch
change request, 2 compilations could happen concurrently
- application restarts are queued. If an application is requested to be
restarted, we have to wait for the previous process to be killed before
starting the new version. We do this by using a lock. We can safely
discard restart requests if there is already a request in queue, because
it would use the latest version in any case
- the test resources service keep alive file should not use the session
start time, because that time changes when using watch mode. This was
trigerring the shutdown of the server even in shared mode.
- the test resources keep alive file was deleted too eagerly. It should
only be deleted IF we know the server will be shutdown. Previously it
was deleted because a request was made, but in the case of watch mode,
other runs will need the service.