Fix file watching (#922)
* Fix file watching

This commit reworks the file watching strategy, in order to make
it easier to understand and to fix some bugs. In particular, the
previous logic was broken in the context of multiproject builds,
depending on whether the build was started from the root project,
or from a submodule. It was for example possible that the whole
project directory was watched, and not just the default sources
and resources directories.

With this change, we removed the separate logic which existed
for whether the user had defined watches or not, to replace it
with a unified version. Tests were added to make sure that it's
possible to watch changes in both single projects and multi-projects.

* Fix file watching under Windows