`stop-testresources-service` should never fail the build
In #1473 the intention was to only fail the build if the user invoked `mvn mn:stop-testresources-service`. This would prevent executions like `mvn mn:run` to fail the build if the Test Resources service was unresponsive.

The problem is that there are more executions that aren't really direct invocations. When the user invokes `mvn test`, the lifecycle is surrounded by a `start-testresources-service` before and a `stop-testresources-service` after. And there is no reliable way to differentiate them from a direct invocation.

Thinking about it, stopping the test resources service is a cleanup operation that is nice to have, but not a must. With this PR, the user will get a clear error message but the build will continue. If another execution is invoked afterwards, the client won't be able to connect to the unresponsive server and a new one will be started. And the unresponsive one should eventually die.